+++ src/ResourceManager/Network/ChangeLog.md
@@ -24,7 +24,17 @@
     - Updated Set-AzureRmVirtualSubnetConfig: Added optional parameter -ServiceEndpoint
 * Added cmdlet to list endpoint services available in the location
     - Get-AzureRmVirtualNetworkAvailableEndpointService
-
+* Added support for SKU parameter to Public IP Addresses and Load Balancers
+    - Updated New-AzureRMLoadBalancer: Added optional parameter -Sku
+    - Updated New-AzureRMPublicIpAddress: Added optional parameter -Sku
+* Added support for DisableOutboundSNAT to Load Balancer Rules
+    - Updated New-AzureRMLoadBalancerRuleConfig: Added optional parameter DisableOutboundSNAT
+    - Updated Add-AzureRMLoadBalancerRuleConfig: Added optional parameter DisableOutboundSNAT
+    - Updated Set-AzureRMLoadBalancerRuleConfig: Added optional parameter DisableOutboundSNAT
+* Added support for IkeV2 P2S
+    - Updated New-AzureRmVirtualNetworkGateway: Added optional parameter -VpnClientProtocol, defaults to [ "SSTP", "IkeV2" ]
+    - Updated Set-AzureRmVirtualNetworkGateway: Added optional parameter -VpnClientProtocol
+	
 ## Version 4.3.1
 
 ## Version 4.3.0